Slår sammen logikk som fikker riktig tagline
