+++ src/components/_common/headers/general-page-header/GeneralPageHeader.tsx
@@ -1,7 +1,7 @@
 import { BodyLong, BodyShort, Heading } from '@navikt/ds-react';
 import {
     PagePropsForPageHeader,
-    getHeaderTagline,
+    getContentTagline,
 } from 'components/_common/headers/sharedHeaderUtils';
 import { Illustration } from 'components/_common/illustration/Illustration';
 import { ContentProps } from 'types/content-props/_content-common';
@@ -16,7 +16,7 @@ type Props = {
 export const GeneralPageHeader = (props: Props) => {
     const { pageProps } = props as { pageProps: PagePropsForPageHeader };
     const illustration = pageProps.data.illustration;
-    const tagLine = getHeaderTagline(pageProps);
+    const tagLine = getContentTagline(pageProps);
     const title = pageProps.data.title || pageProps.displayName;
     const { ingress, hideIngress } = pageProps.data;
