+++ src/components/_common/headers/sharedHeaderUtils.ts
@@ -1,11 +1,6 @@
-import {
-    Audience,
-    ProductDataMixin,
-    getAudience,
-    getSubAudience,
-} from 'types/component-props/_mixins';
+import { ProductDataMixin, getAudience } from 'types/component-props/_mixins';
 import { ContentProps, ContentType } from 'types/content-props/_content-common';
-import { translator } from 'translations';
+import { Language, translator } from 'translations';
 import { getTranslatedTaxonomies, joinWithConjunction } from 'utils/string';
 
 export type PagePropsForPageHeader = ContentProps & {
@@ -15,54 +10,33 @@ export type PagePropsForPageHeader = ContentProps & {
     >;
 };
 
-export const getHeaderTagline = ({ type, language, data }: PagePropsForPageHeader) => {
-    const { taxonomy, audience, customCategory } = data;
+export const getContentTagline = (content: PagePropsForPageHeader, currentLanguage?: Language) => {
+    const { data } = content;
+    const { taxonomy = [], customCategory, audience } = data;
 
-    const currentAudience = getAudience(audience);
-    const subAudience = getSubAudience(audience);
+    const language = currentLanguage || content.language;
+    const selectedAudience = audience && getAudience(audience);
+    const guideTaglines = translator('guides', language);
+    const situationTaglines = translator('situations', language);
 
-    if (currentAudience === Audience.PROVIDER && subAudience && subAudience.length > 0) {
-        const getStringParts = translator('stringParts', language);
-        const getSubAudienceLabel = translator('providerAudience', language);
-        const subAudienceLabels = subAudience.map((audience) =>
-            getSubAudienceLabel(audience).toLowerCase()
-        );
-        return `${getStringParts('for')} ${joinWithConjunction(subAudienceLabels, language)}`;
+    if (content.type === ContentType.GuidePage) {
+        return (selectedAudience && guideTaglines(selectedAudience)) || '';
     }
 
-    if (type === ContentType.SituationPage) {
-        const getTaxonomyLabel = translator('situations', language);
-        return getTaxonomyLabel(currentAudience || 'person');
+    if (content.type === ContentType.SituationPage) {
+        return (selectedAudience && situationTaglines(selectedAudience)) || '';
     }
 
-    if (type === ContentType.GuidePage) {
-        const getTaxonomyLabel = translator('guides', language);
-        return getTaxonomyLabel(currentAudience || 'person');
+    if (taxonomy.length > 0 || customCategory) {
+        const taxonomyStrings = getTranslatedTaxonomies(taxonomy, language);
+        if (customCategory) {
+            taxonomyStrings.push(customCategory);
+        }
+
+        return joinWithConjunction(taxonomyStrings, language);
     }
 
-    if (type === ContentType.Overview) {
-        const getTaxonomyLabel = translator('overview', language);
-        return getTaxonomyLabel('any');
-    }
-
-    if (
-        taxonomy &&
-        (type === ContentType.ThemedArticlePage || type === ContentType.FormIntermediateStepPage)
-    ) {
-        const taxonomyArray = getTranslatedTaxonomies(taxonomy, language);
-        const allCategories = customCategory ? [...taxonomyArray, customCategory] : taxonomyArray;
-
-        return joinWithConjunction(allCategories, language);
-    }
-
-    if (
-        type === ContentType.ProductPage &&
-        (currentAudience === Audience.EMPLOYER || currentAudience === Audience.PROVIDER) // Will catch unlikely events where no sub audience for provider was set
-    ) {
-        const getTaxonomyLabel = translator('products', language);
-        return getTaxonomyLabel(currentAudience);
-    }
-
-    const taxonomyArray = taxonomy ? getTranslatedTaxonomies(taxonomy, language) : [];
-    return joinWithConjunction(taxonomyArray, language);
+    // Catch all if no other taglines could be determined
+    const productAudienceTranslations = translator('products', language);
+    return selectedAudience ? productAudienceTranslations(selectedAudience) : '';
 };
